+++ packages/react/src/components/dialog.tsx
@@ -5,18 +5,18 @@ import {
   forwardRef,
   ReactElement,
   JSXElementConstructor,
-  cloneElement,
   ReactNode,
   createContext,
   useContext,
+  LegacyRef,
 } from "react";
 import { CrossIcon } from "../icons/cross";
 import { clx } from "../utils";
 import { Button } from "./button";
 import { cva } from "class-variance-authority";
+import { Slot } from "@radix-ui/react-slot";
 
 const Dialog = DialogPrimitive.Root;
-type DialogProps = ComponentProps<typeof Dialog>;
 
 /*========================================================================================================================*/
 
@@ -25,8 +25,6 @@ const DialogTrigger: ForwardRefExoticComponent<
 > = forwardRef((props, ref) => {
   return <DialogPrimitive.Trigger ref={ref} {...props} asChild />;
 });
-
-type DialogTriggerProps = ComponentProps<typeof DialogTrigger>;
 
 /*========================================================================================================================*/
 
@@ -44,8 +42,6 @@ const DialogClose: ForwardRefExoticComponent<
     />
   );
 });
-
-type DialogCloseProps = ComponentProps<typeof DialogClose>;
 
 /*========================================================================================================================*/
 
@@ -138,20 +134,17 @@ interface DialogHeaderProps extends ComponentProps<"div"> {
   border?: boolean;
 }
 
-const dialog_header_cva = cva(
-  "flex w-full flex-col text-left px-6 pt-6 pb-4.5",
-  {
-    variants: {
-      border: {
-        true: "border-otl-gray-200 border-b",
-        false: "",
-      },
-    },
-    defaultVariants: {
-      border: false,
+const dialog_header_cva = cva("flex w-full flex-col text-left px-6 pt-6", {
+  variants: {
+    border: {
+      true: "border-otl-gray-200 border-b pb-4.5 mb-6",
+      false: "pb-2",
     },
   },
-);
+  defaultVariants: {
+    border: false,
+  },
+});
 
 const DialogHeader: ForwardRefExoticComponent<DialogHeaderProps> = forwardRef(
   ({ className, border, ...props }, ref) => {
@@ -171,7 +164,7 @@ interface DialogContentProps extends ComponentProps<"div"> {}
 const DialogContent: ForwardRefExoticComponent<DialogContentProps> = forwardRef(
   ({ className, ...props }, ref) => {
     return (
-      <div ref={ref} className={clx("w-full p-6", className)} {...props} />
+      <div ref={ref} className={clx("w-full px-6", className)} {...props} />
     );
   },
 );
@@ -253,7 +246,7 @@ const DialogTitle: ForwardRefExoticComponent<DialogTitleProps> = forwardRef(
     <DialogPrimitive.Title
       ref={ref}
       className={clx(
-        "text-body-lg text-txt-black-900 font-semibold",
+        "text-body-lg text-txt-black-900 font-body font-semibold",
         className,
       )}
       {...props}
@@ -286,7 +279,7 @@ const DialogDescription: ForwardRefExoticComponent<DialogDescriptionProps> =
  * @returns {ReactElement} The cloned child element with the forwarded ref and applied class name.
  */
 
-const dialog_icon_cva = cva("block stroke-[1.5px] size-[30px] shrink-0", {
+const dialog_icon_cva = cva("block stroke-[1.5px] size-10.5 shrink-0", {
   variants: {
     variant: {
       default: "",
@@ -305,14 +298,16 @@ interface DialogIconProps {
   variant: "default" | "primary" | "success" | "warning" | "danger";
   children: ReactElement<any, string | JSXElementConstructor<any>>;
   className?: string;
+  ref?: LegacyRef<HTMLElement>;
 }
 
 const DialogIcon: ForwardRefExoticComponent<DialogIconProps> = forwardRef(
   ({ variant, children, className }, ref) => {
-    return cloneElement(children, {
-      ref,
-      className: clx(dialog_icon_cva({ variant, className })),
-    });
+    return (
+      <Slot ref={ref} className={clx(dialog_icon_cva({ variant }), className)}>
+        {children}
+      </Slot>
+    );
   },
 );
 
@@ -342,16 +337,3 @@ export {
   DialogTitle,
   DialogTrigger,
 };
-export type {
-  DialogProps,
-  DialogTriggerProps,
-  DialogCloseProps,
-  DialogBodyProps,
-  DialogContentProps,
-  DialogHeaderProps,
-  DialogFooterProps,
-  DialogTitleProps,
-  DialogDescriptionProps,
-  DialogIconProps,
-};
-export { dialog_footer_cva };